+++ HealthGuard/Controllers/AdminUserController.cs
@@ -3,13 +3,13 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Authorization;
 using HealthGuard.Services;
-using HealthGuard.Models.DTOs;
+using HealthGuard.Models.Dto; // Đã sửa DTOs -> Dto
 
 namespace HealthGuard.Controllers
 {
     [ApiController]
     [Route("api/admin/users")]
-    // [Authorize(Roles = "ROLE_ADMIN")] // Mở comment để phân quyền Admin
+    // [Authorize(Roles = "ROLE_ADMIN")] 
     public class AdminUserController : ControllerBase
     {
         private readonly AdminUserService _adminUserService;
@@ -30,20 +30,17 @@ namespace HealthGuard.Controllers
         }
 
         [HttpGet("{id}")]
-        public async Task<IActionResult> GetUserByIdAsync([FromRoute] long id)
+        public async Task<IActionResult> GetUserByIdAsync([FromRoute] int id) // Đổi long -> int
         {
             var user = await _adminUserService.GetUserByIdAsync(id);
             return Ok(user);
         }
 
-        // Tương đương với @PatchMapping trong Java
-        // Map<String, Boolean> của Java được chuyển thành Dictionary<string, bool> trong C#
         [HttpPatch("{id}/status")]
         public async Task<IActionResult> UpdateUserStatusAsync(
-            [FromRoute] long id,
+            [FromRoute] int id, // Đổi long -> int
             [FromBody] Dictionary<string, bool> statusUpdate)
         {
-            // Kiểm tra xem key "isActive" có tồn tại trong Dictionary không
             if (!statusUpdate.TryGetValue("isActive", out bool isActive))
             {
                 return BadRequest(new { message = "Thiếu trường isActive trong payload." });
@@ -55,8 +52,8 @@ namespace HealthGuard.Controllers
 
         [HttpPut("{id}/role/{roleId}")]
         public async Task<IActionResult> ChangeUserRoleAsync(
-            [FromRoute] long id,
-            [FromRoute] long roleId)
+            [FromRoute] int id, // Đổi long -> int
+            [FromRoute] int roleId) // Đổi long -> int
         {
             var updatedUser = await _adminUserService.ChangeUserRoleAsync(id, roleId);
             return Ok(updatedUser);
