+++ HealthGuard/Controllers/PatientHistoryController.cs
@@ -7,8 +7,7 @@ namespace HealthGuard.Controllers
 {
     [ApiController]
     [Route("api/patient/history")]
-    [Authorize] // Yêu cầu phải có Token hợp lệ
-    // Bạn có thể dùng [Authorize(Roles = "ROLE_USER")] nếu muốn phân quyền chặt chẽ hơn
+    [Authorize]
     public class PatientHistoryController : ControllerBase
     {
         private readonly PatientHistoryService _historyService;
@@ -23,15 +22,13 @@ namespace HealthGuard.Controllers
             [FromQuery] int page = 0,
             [FromQuery] int size = 10)
         {
-            // Lấy username từ Token (thay thế cho Principal.getName() bên Java)
             string username = User.Identity.Name;
-
             var history = await _historyService.GetMyHistoryListAsync(username, page, size);
             return Ok(history);
         }
 
         [HttpGet("{sessionId}")]
-        public async Task<IActionResult> GetHistoryDetailAsync([FromRoute] long sessionId)
+        public async Task<IActionResult> GetHistoryDetailAsync([FromRoute] int sessionId) // Đổi long -> int
         {
             string username = User.Identity.Name;
             var detail = await _historyService.GetHistoryDetailAsync(username, sessionId);
